extract StoresTree to it's own file
